requirements.txt
streamlit
pandas
numpy
scikit-learn
transformers
torch
sentence-transformers
faiss-cpu
langchain
openai
tiktoken
beautifulsoup4
requests